+++ tests/tools/decompiler_tests.py
@@ -388,6 +388,10 @@ class ParseArgsTests(ToolTestsBase):
         args = parse_args(['decompiler.py', '--compiler-strip', 'file.c'])
         self.assertTrue(args.comp_strip)
 
+    def test_decomp_var_names_is_parsed_correctly_long_form(self):
+        args = parse_args(['decompiler.py', '--var-names', 'simple', 'prog.exe'])
+        self.assertEqual(args.decomp_var_names, 'simple')
+
     def test_generate_archive_is_set_to_false_when_with_archive_not_given(self):
         args = parse_args(['decompiler.py', 'prog.exe'])
         self.assertFalse(args.generate_archive)
@@ -622,6 +626,15 @@ class MainTests(ToolTestsBase):
 
         self.assert_decompilation_was_started_also_with(
             comp_strip=True
+        )
+
+    def test_sets_decomp_var_names_when_given(self):
+        self.call_main_with_standard_arguments_and(
+            '--var-names', 'simple'
+        )
+
+        self.assert_decompilation_was_started_also_with(
+            decomp_var_names='simple'
         )
 
     def test_generates_and_saves_archive_when_requested(self):